Refactor deck views
Update EditDeckView.swift
