+++ Swanki/DecksView.swift
@@ -4,7 +4,7 @@ import SwiftData
 import SwiftUI
 
 /// Show all of the decks in a database.
-struct NewDecksView: View {
+struct DecksView: View {
   @Query private var decks: [Deck]
   @Environment(\.modelContext) private var modelContext
   @State private var editingDeck: Deck?
@@ -40,14 +40,10 @@ struct NewDecksView: View {
       })
     }
     .sheet(item: $editingDeck) { deck in
-      NavigationStack {
-        EditDeckView(deck: deck)
-      }
+      EditDeckView(deck: deck)
     }
     .sheet(isPresented: $shouldShowNewDeck) {
-      NavigationStack {
-        EditDeckView(deck: nil)
-      }
+      EditDeckView(deck: nil)
     }
     .navigationTitle("Decks")
     .fileImporter(isPresented: $shouldShowFileImporter, allowedContentTypes: [.ankiPackage], onCompletion: { result in
@@ -89,51 +85,9 @@ struct NewDecksView: View {
   }
 }
 
-struct EditDeckView: View {
-  var deck: Deck?
-
-  @State private var name: String = ""
-  @Environment(\.dismiss) private var dismiss
-  @Environment(\.modelContext) private var modelContext
-
-  var body: some View {
-    Form {
-      TextField("Name", text: $name)
-    }
-    .navigationTitle(deck == nil ? "New Deck" : "Edit Deck")
-    .toolbar {
-      ToolbarItem(placement: .confirmationAction) {
-        Button {
-          if let deck {
-            deck.name = name
-          } else {
-            let deck = Deck(name: name)
-            modelContext.insert(deck)
-          }
-          dismiss()
-        } label: {
-          Text("Done")
-        }
-      }
-      ToolbarItem(placement: .cancellationAction) {
-        Button {
-          dismiss()
-        } label: {
-          Text("Cancel")
-        }
-      }
-    }
-    .onAppear {
-      if let deck {
-        name = deck.name
-      }
-    }
-  }
-}
-
 #Preview {
   NavigationStack {
-    NewDecksView()
+    DecksView()
   }
   .modelContainer(.previews)
 }
